// front/src/app/api-client/model/reminder-request.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecurrenceRuleRequest } from './recurrence-rule-request';


/**
 * Création / modification d\'un reminder. Les AgendaEntry sont générées automatiquement.
 */
export interface ReminderRequest { 
    titre: string;
    description?: string;
    type: ReminderRequest.TypeEnum;
    /**
     * Début de la 1re occurrence
     */
    dateHeureDebut: string;
    /**
     * Fin (optionnelle) de la 1re occurrence ; sert à calculer la durée
     */
    dateHeureFin?: string;
    /**
     * Identifiants des membres concernés
     */
    membreIds?: Array<number>;
    /**
     * Reminder récurrent ? Si vrai, \'recurrence\' est obligatoire.
     */
    isRecurring?: boolean;
    recurrence?: RecurrenceRuleRequest;
}
export namespace ReminderRequest {
    export const TypeEnum = {
        Sport: 'SPORT',
        TestMedical: 'TEST_MEDICAL',
        RendezVous: 'RENDEZ_VOUS',
        Autre: 'AUTRE',
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}